+++ Editor/ShaderGUI/MaterialGUI.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
+using UnityEngine.Assertions;
 
 namespace UniToon
 {
@@ -64,4 +65,65 @@ namespace UniToon
             return res;
         }
     }
+
+    class SavedParameter<T>
+            where T : IEquatable<T>
+        {
+            internal delegate void SetParameter(string key, T value);
+            internal delegate T GetParameter(string key, T defaultValue);
+
+            readonly string m_Key;
+            bool m_Loaded;
+            T m_Value;
+
+            readonly SetParameter m_Setter;
+            readonly GetParameter m_Getter;
+
+            public SavedParameter(string key, T value, GetParameter getter, SetParameter setter)
+            {
+                Assert.IsNotNull(setter);
+                Assert.IsNotNull(getter);
+
+                m_Key = key;
+                m_Loaded = false;
+                m_Value = value;
+                m_Setter = setter;
+                m_Getter = getter;
+            }
+
+            void Load()
+            {
+                if (m_Loaded)
+                    return;
+
+                m_Loaded = true;
+                m_Value = m_Getter(m_Key, m_Value);
+            }
+
+            public T value
+            {
+                get
+                {
+                    Load();
+                    return m_Value;
+                }
+                set
+                {
+                    Load();
+
+                    if (m_Value.Equals(value))
+                        return;
+
+                    m_Value = value;
+                    m_Setter(m_Key, value);
+                }
+            }
+        }
+
+        // Pre-specialized class for easier use and compatibility with existing code
+        sealed class SavedBool : SavedParameter<bool>
+        {
+            public SavedBool(string key, bool value)
+                : base(key, value, EditorPrefs.GetBool, EditorPrefs.SetBool) { }
+        }
 }